updating my fork

+++ config/file.go
@@ -1,36 +1,35 @@
 // Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
-// See License.txt for license information.
+// See LICENSE.txt for license information.
 
 package config
 
 import (
-	"bytes"
 	"fmt"
-	"io"
 	"io/ioutil"
 	"os"
 	"path/filepath"
 
 	"github.com/pkg/errors"
 
-	"github.com/mattermost/mattermost-server/mlog"
-	"github.com/mattermost/mattermost-server/model"
-	"github.com/mattermost/mattermost-server/utils/fileutils"
+	"github.com/mattermost/mattermost-server/v5/model"
+	"github.com/mattermost/mattermost-server/v5/shared/mlog"
+	"github.com/mattermost/mattermost-server/v5/utils/fileutils"
 )
 
 var (
+	// ErrReadOnlyConfiguration is returned when an attempt to modify a read-only configuration is made.
 	ErrReadOnlyConfiguration = errors.New("configuration is read-only")
 )
 
 // FileStore is a config store backed by a file such as config/config.json.
 //
 // It also uses the folder containing the configuration file for storing other configuration files.
+// Not to be used directly. Only to be used as a backing store for config.Store
 type FileStore struct {
-	commonStore
-
-	path    string
-	watch   bool
-	watcher *watcher
+	path     string
+	watch    bool
+	watcher  *watcher
+	callback func()
 }
 
 // NewFileStore creates a new instance of a config store backed by the given file path.
@@ -45,15 +44,6 @@ func NewFileStore(path string, watch bool) (fs *FileStore, err error) {
 	fs = &FileStore{
 		path:  resolvedPath,
 		watch: watch,
-	}
-	if err = fs.Load(); err != nil {
-		return nil, errors.Wrap(err, "failed to load")
-	}
-
-	if fs.watch {
-		if err = fs.startWatcher(); err != nil {
-			mlog.Error("failed to start config watcher", mlog.String("path", path), mlog.Err(err))
-		}
 	}
 
 	return fs, nil
@@ -92,33 +82,45 @@ func resolveConfigFilePath(path string) (string, error) {
 	return "", fmt.Errorf("failed to find config file %s", path)
 }
 
-// Set replaces the current configuration in its entirety and updates the backing store.
-func (fs *FileStore) Set(newCfg *model.Config) (*model.Config, error) {
-	return fs.commonStore.set(newCfg, func(cfg *model.Config) error {
-		if *fs.config.ClusterSettings.Enable && *fs.config.ClusterSettings.ReadOnlyConfig {
-			return ErrReadOnlyConfiguration
-		}
+// resolveFilePath uses the name if name is absolute path.
+// otherwise returns the combined path/name
+func (fs *FileStore) resolveFilePath(name string) string {
+	// Absolute paths are explicit and require no resolution.
+	if filepath.IsAbs(name) {
+		return name
+	}
+	return filepath.Join(filepath.Dir(fs.path), name)
+}
 
-		return fs.commonStore.validate(cfg)
-	}, fs.persist)
+// Set replaces the current configuration in its entirety and updates the backing store.
+func (fs *FileStore) Set(newCfg *model.Config) error {
+	if *newCfg.ClusterSettings.Enable && *newCfg.ClusterSettings.ReadOnlyConfig {
+		return ErrReadOnlyConfiguration
+	}
+
+	return fs.persist(newCfg)
 }
 
 // persist writes the configuration to the configured file.
 func (fs *FileStore) persist(cfg *model.Config) error {
-	fs.stopWatcher()
+	needsRestart := false
+	if fs.watcher != nil {
+		fs.stopWatcher()
+		needsRestart = true
+	}
 
 	b, err := marshalConfig(cfg)
 	if err != nil {
 		return errors.Wrap(err, "failed to serialize")
 	}
 
-	err = ioutil.WriteFile(fs.path, b, 0644)
+	err = ioutil.WriteFile(fs.path, b, 0600)
 	if err != nil {
 		return errors.Wrap(err, "failed to write file")
 	}
 
-	if fs.watch {
-		if err = fs.startWatcher(); err != nil {
+	if fs.watch && needsRestart {
+		if err = fs.Watch(fs.callback); err != nil {
 			mlog.Error("failed to start config watcher", mlog.String("path", fs.path), mlog.Err(err))
 		}
 	}
@@ -127,40 +129,27 @@ func (fs *FileStore) persist(cfg *model.Config) error {
 }
 
 // Load updates the current configuration from the backing store.
-func (fs *FileStore) Load() (err error) {
-	var needsSave bool
-	var f io.ReadCloser
-
-	f, err = os.Open(fs.path)
+func (fs *FileStore) Load() ([]byte, error) {
+	f, err := os.Open(fs.path)
 	if os.IsNotExist(err) {
-		needsSave = true
-		defaultCfg := &model.Config{}
-		defaultCfg.SetDefaults()
-
-		var defaultCfgBytes []byte
-		defaultCfgBytes, err = marshalConfig(defaultCfg)
-		if err != nil {
-			return errors.Wrap(err, "failed to serialize default config")
-		}
-
-		f = ioutil.NopCloser(bytes.NewReader(defaultCfgBytes))
+		return nil, nil
 
 	} else if err != nil {
-		return errors.Wrapf(err, "failed to open %s for reading", fs.path)
+		return nil, errors.Wrapf(err, "failed to open %s for reading", fs.path)
 	}
-	defer func() {
-		closeErr := f.Close()
-		if err == nil && closeErr != nil {
-			err = errors.Wrap(closeErr, "failed to close")
-		}
-	}()
+	defer f.Close()
 
-	return fs.commonStore.load(f, needsSave, fs.commonStore.validate, fs.persist)
+	fileBytes, err := ioutil.ReadAll(f)
+	if err != nil {
+		return nil, err
+	}
+
+	return fileBytes, nil
 }
 
 // GetFile fetches the contents of a previously persisted configuration file.
 func (fs *FileStore) GetFile(name string) ([]byte, error) {
-	resolvedPath := filepath.Join(filepath.Dir(fs.path), name)
+	resolvedPath := fs.resolveFilePath(name)
 
 	data, err := ioutil.ReadFile(resolvedPath)
 	if err != nil {
@@ -170,11 +159,17 @@ func (fs *FileStore) GetFile(name string) ([]byte, error) {
 	return data, nil
 }
 
+// GetFilePath returns the resolved path of a configuration file.
+// The file may not necessarily exist.
+func (fs *FileStore) GetFilePath(name string) string {
+	return fs.resolveFilePath(name)
+}
+
 // SetFile sets or replaces the contents of a configuration file.
 func (fs *FileStore) SetFile(name string, data []byte) error {
-	resolvedPath := filepath.Join(filepath.Dir(fs.path), name)
+	resolvedPath := fs.resolveFilePath(name)
 
-	err := ioutil.WriteFile(resolvedPath, data, 0777)
+	err := ioutil.WriteFile(resolvedPath, data, 0600)
 	if err != nil {
 		return errors.Wrapf(err, "failed to write file to %s", resolvedPath)
 	}
@@ -188,7 +183,7 @@ func (fs *FileStore) HasFile(name string) (bool, error) {
 		return false, nil
 	}
 
-	resolvedPath := filepath.Join(filepath.Dir(fs.path), name)
+	resolvedPath := fs.resolveFilePath(name)
 
 	_, err := os.Stat(resolvedPath)
 	if err != nil && os.IsNotExist(err) {
@@ -202,6 +197,11 @@ func (fs *FileStore) HasFile(name string) (bool, error) {
 
 // RemoveFile removes a previously persisted configuration file.
 func (fs *FileStore) RemoveFile(name string) error {
+	if filepath.IsAbs(name) {
+		// Don't delete absolute filenames, as may be mounted drive, etc.
+		mlog.Debug("Skipping removal of configuration file with absolute path", mlog.String("filename", name))
+		return nil
+	}
 	resolvedPath := filepath.Join(filepath.Dir(fs.path), name)
 
 	err := os.Remove(resolvedPath)
@@ -212,20 +212,16 @@ func (fs *FileStore) RemoveFile(name string) error {
 		return errors.Wrap(err, "failed to remove file")
 	}
 
-	return err
+	return nil
 }
 
-// startWatcher starts a watcher to monitor for external config file changes.
-func (fs *FileStore) startWatcher() error {
-	if fs.watcher != nil {
+func (fs *FileStore) Watch(callback func()) error {
+	if fs.watcher != nil || !fs.watch {
 		return nil
 	}
 
-	watcher, err := newWatcher(fs.path, func() {
-		if err := fs.Load(); err != nil {
-			mlog.Error("failed to reload file on change", mlog.String("path", fs.path), mlog.Err(err))
-		}
-	})
+	fs.callback = callback
+	watcher, err := newWatcher(fs.path, callback)
 	if err != nil {
 		return err
 	}
@@ -254,9 +250,6 @@ func (fs *FileStore) String() string {
 
 // Close cleans up resources associated with the store.
 func (fs *FileStore) Close() error {
-	fs.configLock.Lock()
-	defer fs.configLock.Unlock()
-
 	fs.stopWatcher()
 
 	return nil
